requirements.txt
asgiref==3.2.3
beautifulsoup4==4.8.2
bs4==0.0.1
certifi==2019.11.28
chardet==3.0.4
Django==3.0.4
djangorestframework==3.11.0
idna==2.9
MarkupSafe==1.1.1
newsapi-python==0.2.6
Pillow==7.0.0
pytz==2019.3
requests==2.23.0
soupsieve==2.0
sqlparse==0.3.1
urllib3==1.25.8
